+++ modules/nextflow/src/main/groovy/nextflow/k8s/K8sTaskHandler.groovy
@@ -223,9 +223,6 @@ class K8sTaskHandler extends TaskHandler implements FusionAwareTask {
         if (System.getenv("LATCH_NF_DEBUG") != "true") {
             def execId = System.getenv("FLYTE_INTERNAL_EXECUTION_ID")
             builder.withEnv(PodEnv.value("FLYTE_INTERNAL_EXECUTION_ID", execId))
-
-            def logDir = System.getenv("LATCH_LOG_DIR")
-            builder.withEnv(PodEnv.value("LATCH_LOG_DIR", logDir))
         }
 
         // when `entrypointOverride` is false the launcher is run via `args` instead of `command`