+++ modules/nextflow/src/main/groovy/nextflow/processor/PublishDir.groovy
@@ -383,21 +383,21 @@ class PublishDir {
         try {
             processFileImpl(source, destination)
         }
-        catch( FileAlreadyExistsException e ) {
-            if( checkIsSameRealPath(source, destination) )
+        catch ( FileAlreadyExistsException e ) {
+            if ( checkIsSameRealPath(source, destination) )
                 return 
             // make sure destination and source does not overlap
             // see https://github.com/nextflow-io/nextflow/issues/2177
-            if( checkSourcePathConflicts(destination))
+            if ( checkSourcePathConflicts(destination))
                 return
             
-            if( overwrite ) {
+            if ( overwrite ) {
                 log.warn "Overwriting file at ${destination.toUriString()}"
                 FileHelper.deletePath(destination)
                 processFileImpl(source, destination)
+            } else {
+                log.debug "Skipping file. File already exists at ${destination.toUriString()}"
             }
-
-            log.debug "Skipping file. File already exists at ${destination.toUriString()}"
         }
 
         notifyFilePublish(destination, source)