+++ modules/nextflow/src/main/groovy/nextflow/processor/TaskProcessor.groovy
@@ -15,6 +15,7 @@
  */
 package nextflow.processor
 
+import java.nio.file.StandardCopyOption
 
 import static nextflow.processor.ErrorStrategy.*
 
@@ -2374,6 +2375,45 @@ class TaskProcessor {
         isCacheable() && session.resumeMode
     }
 
+    private void uploadTaskLogs( TaskRun task ) {
+        def logDir = System.getenv("LATCH_LOG_DIR")
+        if (logDir == null) {
+            return
+        }
+
+        Path p = FileHelper.asPath(logDir)
+        if (p.scheme != 'latch') {
+            log.warn "LATCH_LOG_DIR ${logDir} is not a valid latch directory"
+            return
+        }
+
+        log.debug "Uploading log files for ${safeTaskName(task)}"
+
+        for (String name : [
+            TaskRun.CMD_LOG,
+            TaskRun.CMD_SCRIPT,
+            TaskRun.CMD_INFILE,
+            TaskRun.CMD_OUTFILE,
+            TaskRun.CMD_ERRFILE,
+            TaskRun.CMD_EXIT,
+            TaskRun.CMD_START,
+            TaskRun.CMD_RUN,
+            TaskRun.CMD_STAGE,
+            TaskRun.CMD_TRACE,
+            TaskRun.CMD_ENV
+        ]) {
+            try {
+                Path source = task.workDir.resolve(name)
+                Path subPath = session.workDir.relativize(source)
+                Path target = p.resolve("work").resolve(subPath)
+
+                FileHelper.copyPath(source, target, StandardCopyOption.REPLACE_EXISTING)
+            } catch (NoSuchFileException e) {
+                log.debug "Failed to upload ${name} for ${safeTaskName(task)}: ${e.toString()}"
+            }
+        }
+    }
+
     /**
      * Finalize the task execution, checking the exit status
      * and binding output values accordingly
@@ -2383,6 +2423,12 @@ class TaskProcessor {
      */
     private void finalizeTask0( TaskRun task ) {
         log.trace "Finalize process > ${safeTaskName(task)}"
+
+        try {
+            uploadTaskLogs(task)
+        } catch (Exception e) {
+            log.warn "Failed to upload log files for ${safeTaskName(task)}: ${e.toString()}\n${e.printStackTrace()}"
+        }
 
         // -- bind output (files)
         if( task.canBind ) {